Interfaces moved to separate folder, updated imports of interfaces
