+++ src/interfaces/IAlternativeApi.ts
