+++ src/interfaces/IWeatherData.ts
